TIM-453: fix TD all-black render — LP64 CRC, MixFile bsearch, iconset layout
Three LP64 bugs prevented terrain tiles from rendering on Linux:

1. MiscAsm.cpp Calculate_CRC: used `unsigned long` for the rotating-add
   CRC accumulator.  On LP64 the 64-bit `>> 31` shift gives a different
   result than the 32-bit Windows original, so every MixFile hash
   mismatched and MixFileClass::Retrieve returned NULL for all cached
   files.  Fixed by using `uint32_t` throughout.

2. MIXFILE.CPP compfunc: compared SubBlock entries as `long` (8 bytes),
   reading `CRC` + `Offset` together as a 64-bit value instead of just
   the 4-byte CRC field.  Fixed by casting to `int32_t const *`.

3. td-win32-stubs.cpp Buffer_Draw_Stamp / Buffer_Draw_Stamp_Clip: read
   the IControl_Type header via the host-struct pointer, which on LP64
   places `Icons` at offset 24 instead of the on-disk offset 12.  Fixed
   by defining IControl_Disk (packed int32_t layout matching Win32) and
   treating the `Icons` field as a byte-offset from the icondata base.

Result: 165/166 desert terrain templates load, 38 distinct palette colors
visible at frames 50/100/150 (dominant sandy brown #885844), full 200-frame
TD_AUTOSTART+TD_CHEAT run completes without crash.

Also adds TD_AUTOSTART / TD_CHEAT audit system (init milestones,
screenshot capture at frames 50/100/150, flag-to-win at frame 200).

+++ TIBERIANDAWN/MiscAsm.cpp
@@ -597,28 +597,30 @@ void __cdecl Fat_Put_Pixel(int x, int y, int color, int siz, GraphicViewPortClas
 extern "C" long __cdecl Calculate_CRC(void *buffer, long length)
 {
 #ifndef _MSC_VER
-	// TIM-339: rotating additive CRC — for each 4-byte word: crc=ROL(crc,1)+word;
-	// remainder bytes are assembled little-endian before the final accumulate step.
-	unsigned long crc = 0;
+	/* TIM-453: rotating additive CRC — use uint32_t throughout so the ROL(crc,1)
+	 * rotation is 32-bit, matching the Win32 assembly original.  `unsigned long`
+	 * is 64 bits on LP64 Linux, making `crc >> 31` a 64-bit shift and producing
+	 * wrong hash values that fail MixFile bsearch lookups. */
+	uint32_t crc = 0;
 	const unsigned char *src = (const unsigned char*)buffer;
-	unsigned long len = (unsigned long)length;
-	unsigned long full_words = len >> 2;
-	unsigned long remainder  = len & 3;
+	uint32_t len = (uint32_t)length;
+	uint32_t full_words = len >> 2;
+	uint32_t remainder  = len & 3;
 
-	for (unsigned long i = 0; i < full_words; i++, src += 4) {
-		unsigned long word;
+	for (uint32_t i = 0; i < full_words; i++, src += 4) {
+		uint32_t word;
 		memcpy(&word, src, 4);  // little-endian load (x86 native order)
 		crc = (crc << 1) | (crc >> 31);
 		crc += word;
 	}
 	if (remainder) {
-		unsigned long word = 0;
-		for (unsigned long i = 0; i < remainder; i++)
-			word |= ((unsigned long)src[i]) << (i * 8);
+		uint32_t word = 0;
+		for (uint32_t i = 0; i < remainder; i++)
+			word |= ((uint32_t)src[i]) << (i * 8);
 		crc = (crc << 1) | (crc >> 31);
 		crc += word;
 	}
-	return (long)crc;
+	return (long)(int32_t)crc;
 #else
 	{}
 #endif
